Weather Forecast Application
This Application is created with an aim to display the weather conditions of different countries and its various cities.
The key features of the application includes displaying the icons according to the weather condition at a place.
The second key features is  the user can change the units in which he or she wants to get the details about the weather at his or her place.